requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
pyyaml>=6.0
loguru>=0.7.0
openpyxl>=3.1.0
scipy>=1.11.0
statsmodels>=0.14.0
